requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
dj-rest-auth==4.0.1
Django==4.2.4
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
phonenumberslite==8.13.18
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
